+++ src/main/java/javax/util/streamex/DistinctSpliterator.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2015 Tagir Valeev
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package javax.util.streamex;
 
 import java.util.Comparator;
@@ -8,12 +23,12 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Consumer;
 
-final class DistinctSpliterator<T> implements Spliterator<T> {
-    final Spliterator<T> source;
-    AtomicLong nullCounter;
-    Map<T, Long> counts;
-    final long atLeast;
-    T cur;
+/* package */ final class DistinctSpliterator<T> implements Spliterator<T> {
+    private final Spliterator<T> source;
+    private AtomicLong nullCounter;
+    private Map<T, Long> counts;
+    private final long atLeast;
+    private T cur;
 
     DistinctSpliterator(Spliterator<T> source, long atLeast, AtomicLong nullCounter, Map<T, Long> counts) {
         this.source = source;