+++ src/main/java/javax/util/streamex/IntStreamEx.java
@@ -25,6 +25,7 @@ import java.util.List;
 import java.util.OptionalDouble;
 import java.util.OptionalInt;
 import java.util.Random;
+import java.util.Map.Entry;
 import java.util.PrimitiveIterator.OfInt;
 import java.util.concurrent.ForkJoinPool;
 import java.util.function.BiConsumer;
@@ -198,6 +199,25 @@ public class IntStreamEx implements IntStream {
         return strategy().newDoubleStreamEx(stream.mapToDouble(mapper));
     }
     
+    /**
+     * Returns an {@link EntryStream} consisting of the {@link Entry} objects
+     * which keys and values are results of applying the given functions to the
+     * elements of this stream.
+     *
+     * <p>
+     * This is an intermediate operation.
+     *
+     * @param <K>
+     *            The {@code Entry} key type
+     * @param <V>
+     *            The {@code Entry} value type
+     * @param keyMapper
+     *            a non-interfering, stateless function to apply to each element
+     * @param valueMapper
+     *            a non-interfering, stateless function to apply to each element
+     * @return the new stream
+     * @since 0.3.1
+     */
     public <K, V> EntryStream<K, V> mapToEntry(IntFunction<? extends K> keyMapper, IntFunction<? extends V> valueMapper) {
         return strategy().newEntryStream(
                 stream.mapToObj(t -> new AbstractMap.SimpleImmutableEntry<>(keyMapper.apply(t), valueMapper.apply(t))));
@@ -1048,7 +1068,7 @@ public class IntStreamEx implements IntStream {
 
     /**
      * Returns a {@code char[]} array containing the elements of this stream
-     * which are converted to bytes using {@code (char)} cast operation.
+     * which are converted to chars using {@code (char)} cast operation.
      *
      * <p>
      * This is a terminal operation.
@@ -1063,7 +1083,7 @@ public class IntStreamEx implements IntStream {
 
     /**
      * Returns a {@code short[]} array containing the elements of this stream
-     * which are converted to bytes using {@code (short)} cast operation.
+     * which are converted to shorts using {@code (short)} cast operation.
      *
      * <p>
      * This is a terminal operation.